Remove old files, refactor
