+++ src/app/quiz/components/QuizIntro.tsx
@@ -1,0 +1,140 @@
+'use client';
+
+import { ChevronRight, User, Users } from 'lucide-react';
+import { motion } from 'motion/react';
+
+import { usePCTheme } from '@/hooks/usePCTheme';
+
+interface QuizIntroProps {
+  onStartSolo: () => void;
+  onStartGroup: () => void;
+}
+
+export function QuizIntro({ onStartSolo, onStartGroup }: QuizIntroProps) {
+  const { isDark } = usePCTheme();
+
+  return (
+    <div className="flex-1 flex flex-col items-center justify-center px-5 py-12 min-h-[80vh]">
+      <motion.div
+        initial={{ opacity: 0, y: 30 }}
+        animate={{ opacity: 1, y: 0 }}
+        transition={{ duration: 0.5 }}
+        className="w-full max-w-md"
+      >
+        <div className="text-center mb-10">
+          <div className="text-4xl mb-4">🍿</div>
+          <h1
+            className="mb-2"
+            style={{
+              fontFamily: "var(--font-bebas-neue), 'Bebas Neue', sans-serif",
+              fontSize: '2.2rem',
+              letterSpacing: '0.05em',
+              color: 'var(--pc-t1)',
+            }}
+          >
+            Let&apos;s find your movie
+          </h1>
+          <p style={{ color: 'var(--pc-t3)', fontSize: '0.9rem' }}>
+            Are you watching solo or with others?
+          </p>
+        </div>
+
+        <div className="flex flex-col gap-4">
+          {/* Solo */}
+          <button
+            onClick={onStartSolo}
+            className="group flex items-center gap-5 p-5 rounded-2xl text-left transition-all duration-200 active:scale-[0.98]"
+            style={{
+              background: 'var(--pc-surface)',
+              border: '1px solid var(--pc-bd2)',
+            }}
+            onMouseEnter={(e) => {
+              (e.currentTarget as HTMLElement).style.borderColor = isDark
+                ? 'rgba(245,197,24,0.4)'
+                : 'rgba(196,149,10,0.4)';
+              (e.currentTarget as HTMLElement).style.background = 'var(--pc-surface-hover)';
+            }}
+            onMouseLeave={(e) => {
+              (e.currentTarget as HTMLElement).style.borderColor = 'var(--pc-bd2)';
+              (e.currentTarget as HTMLElement).style.background = 'var(--pc-surface)';
+            }}
+          >
+            <div
+              className="w-12 h-12 rounded-xl flex items-center justify-center shrink-0"
+              style={{
+                background: 'rgba(245,197,24,0.15)',
+                color: 'var(--pc-gold)',
+              }}
+            >
+              <User size={22} />
+            </div>
+            <div>
+              <div
+                style={{
+                  color: 'var(--pc-t1)',
+                  fontWeight: 600,
+                  fontSize: '1rem',
+                }}
+              >
+                Just me
+              </div>
+              <div style={{ color: 'var(--pc-t4)', fontSize: '0.85rem' }}>
+                Solo movie night — personalized just for you
+              </div>
+            </div>
+            <ChevronRight
+              size={18}
+              className="ml-auto shrink-0"
+              style={{ color: 'var(--pc-t4)' }}
+            />
+          </button>
+
+          {/* Group */}
+          <button
+            onClick={onStartGroup}
+            className="group flex items-center gap-5 p-5 rounded-2xl text-left transition-all duration-200 active:scale-[0.98]"
+            style={{
+              background: 'var(--pc-surface)',
+              border: '1px solid var(--pc-bd2)',
+            }}
+            onMouseEnter={(e) => {
+              (e.currentTarget as HTMLElement).style.borderColor = 'rgba(139,92,246,0.4)';
+              (e.currentTarget as HTMLElement).style.background = 'var(--pc-surface-hover)';
+            }}
+            onMouseLeave={(e) => {
+              (e.currentTarget as HTMLElement).style.borderColor = 'var(--pc-bd2)';
+              (e.currentTarget as HTMLElement).style.background = 'var(--pc-surface)';
+            }}
+          >
+            <div
+              className="w-12 h-12 rounded-xl flex items-center justify-center shrink-0"
+              style={{ background: 'rgba(139,92,246,0.15)', color: '#8B5CF6' }}
+            >
+              <Users size={22} />
+            </div>
+            <div>
+              <div
+                style={{
+                  color: 'var(--pc-t1)',
+                  fontWeight: 600,
+                  fontSize: '1rem',
+                }}
+              >
+                Group mode 🎉
+              </div>
+              <div style={{ color: 'var(--pc-t4)', fontSize: '0.85rem' }}>
+                Find a film everyone will enjoy
+              </div>
+            </div>
+            <div
+              className="ml-auto shrink-0 text-xs px-2 py-1 rounded-full"
+              style={{ background: 'rgba(139,92,246,0.15)', color: '#A78BFA' }}
+            >
+              NEW
+            </div>
+          </button>
+        </div>
+      </motion.div>
+    </div>
+  );
+}
