+++ src/app/quiz/components/index.ts
@@ -1,0 +1,5 @@
+export { BetweenPersons } from './BetweenPersons';
+export { GroupSetup } from './GroupSetup';
+export { QuizIntro } from './QuizIntro';
+export { QuizNavigation } from './QuizNavigation';
+export { EraStep, FavoriteActorStep, FavoriteMovieStep, MoodStep, ToneStep } from './steps';
