+++ src/app/quiz/components/steps/index.ts
@@ -1,0 +1,5 @@
+export { EraStep } from './EraStep';
+export { FavoriteActorStep } from './FavoriteActorStep';
+export { FavoriteMovieStep } from './FavoriteMovieStep';
+export { MoodStep } from './MoodStep';
+export { ToneStep } from './ToneStep';
