+++ src/app/quiz/types.ts
@@ -1,0 +1,14 @@
+export type Era = 'new' | 'classic' | 'both';
+export type Tone = 'light' | 'balanced' | 'serious' | 'dark';
+
+export interface PersonAnswers {
+  name: string;
+  favoriteMovie: string;
+  favoriteMovieWhy: string;
+  era: Era | '';
+  moods: string[];
+  tone: Tone | '';
+  favoriteActor: string;
+}
+
+export type Phase = 'intro' | 'group-setup' | 'questions' | 'between-persons';
