+++ src/app/results/components/index.ts
@@ -1,0 +1,6 @@
+export { ExpandedSuggestion } from './ExpandedSuggestion';
+export { GenrePill } from './GenrePill';
+export { MainMovieCard } from './MainMovieCard';
+export { SimilarityBadge } from './SimilarityBadge';
+export { SmallSuggestionCard } from './SmallSuggestionCard';
+export { StarRating } from './StarRating';
