+++ src/app/results/page.tsx
@@ -1,21 +1,14 @@
 'use client';
 
-import {
-  ChevronLeft,
-  ChevronRight,
-  Clapperboard,
-  Clock,
-  RotateCcw,
-  Sparkles,
-  Star,
-  Users,
-} from 'lucide-react';
+import { ChevronLeft, ChevronRight, RotateCcw, Sparkles, Users } from 'lucide-react';
 import { AnimatePresence, motion } from 'motion/react';
 import { useRouter } from 'next/navigation';
 import { useCallback, useEffect, useRef, useState } from 'react';
 
 import { usePCTheme } from '@/hooks/usePCTheme';
 import { enhanceMoviesWithPosters, type MovieRecommendation } from '@/utils/client';
+
+import { ExpandedSuggestion, MainMovieCard, SmallSuggestionCard } from './components';
 
 interface ApiResponse {
   title: string;
@@ -40,482 +33,6 @@ interface ApiResponse {
     aiDescription?: string;
     isMainRecommendation?: boolean;
   }[];
-}
-
-function StarRating({ score }: { score: number }) {
-  const stars = 5;
-  const filled = Math.round((score / 10) * stars);
-  return (
-    <div className="flex items-center gap-0.5">
-      {Array.from({ length: stars }).map((_, i) => (
-        <Star
-          key={i}
-          size={12}
-          fill={i < filled ? '#F5C518' : 'none'}
-          stroke={i < filled ? '#F5C518' : 'var(--pc-t4)'}
-        />
-      ))}
-    </div>
-  );
-}
-
-function SimilarityBadge({ similarity }: { similarity: number }) {
-  const pct = Math.round(similarity * 100);
-  const color = pct >= 95 ? '#14B8A6' : pct >= 90 ? '#F5C518' : pct >= 85 ? '#FF9F1C' : '#8B5CF6';
-  return (
-    <div
-      className="flex items-center gap-1.5 px-2.5 py-1 rounded-full text-xs"
-      style={{
-        background: `${color}18`,
-        border: `1px solid ${color}35`,
-        color,
-      }}
-    >
-      <Sparkles size={10} />
-      {pct}% match
-    </div>
-  );
-}
-
-function GenrePill({ genre }: { genre: string }) {
-  return (
-    <span
-      className="px-2.5 py-1 rounded-full text-xs"
-      style={{
-        background: 'var(--pc-bd1)',
-        border: '1px solid var(--pc-bd2)',
-        color: 'var(--pc-t2)',
-      }}
-    >
-      {genre}
-    </span>
-  );
-}
-
-function MainMovieCard({ movie }: { movie: MovieRecommendation }) {
-  const { isDark } = usePCTheme();
-  const [imgLoaded, setImgLoaded] = useState(false);
-  const score = movie.score_rating ?? 0;
-  const pct = Math.round(movie.similarity * 100);
-
-  return (
-    <motion.div
-      initial={{ opacity: 0, y: 30 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ duration: 0.7, delay: 0.1 }}
-      className="relative overflow-hidden rounded-3xl"
-      style={{
-        background: 'var(--pc-surface)',
-        border: '1px solid var(--pc-bd2)',
-        boxShadow: 'var(--pc-card-shadow)',
-      }}
-    >
-      {movie.posterURL && (
-        <div className="relative h-72 md:h-96 overflow-hidden">
-          <div
-            className="absolute inset-0 transition-opacity duration-700"
-            style={{
-              opacity: imgLoaded ? 0 : 1,
-              background: 'var(--pc-surface-deep)',
-            }}
-          />
-          <img
-            src={movie.posterURL}
-            alt={movie.name}
-            className="w-full h-full object-cover transition-opacity duration-700"
-            style={{ opacity: imgLoaded ? 1 : 0 }}
-            onLoad={() => setImgLoaded(true)}
-          />
-          <div
-            className="absolute inset-0"
-            style={{
-              background: isDark
-                ? 'linear-gradient(to bottom, rgba(9,9,15,0.2) 0%, rgba(9,9,15,0.7) 60%, #13131F 100%)'
-                : 'linear-gradient(to bottom, rgba(255,255,255,0.1) 0%, rgba(255,255,255,0.6) 60%, #FFFFFF 100%)',
-            }}
-          />
-
-          {/* Match badge */}
-          <div className="absolute top-4 right-4">
-            <SimilarityBadge similarity={movie.similarity} />
-          </div>
-
-          {/* AI pick label */}
-          <div className="absolute top-4 left-4">
-            <div
-              className="flex items-center gap-1.5 px-3 py-1.5 rounded-full text-xs"
-              style={{
-                background: isDark ? 'rgba(9,9,15,0.8)' : 'rgba(255,255,255,0.85)',
-                border: '1px solid rgba(245,197,24,0.25)',
-                color: 'var(--pc-gold)',
-                backdropFilter: 'blur(8px)',
-              }}
-            >
-              <Sparkles size={10} />
-              AI Pick
-            </div>
-          </div>
-
-          {/* Title overlay */}
-          <div className="absolute bottom-6 left-6 right-6">
-            <h2
-              className="mb-1"
-              style={{
-                fontFamily: "var(--font-bebas-neue), 'Bebas Neue', sans-serif",
-                fontSize: 'clamp(1.8rem, 5vw, 3rem)',
-                letterSpacing: '0.04em',
-                color: isDark ? '#FFFFFF' : '#0D0D1A',
-                lineHeight: 1,
-                textShadow: isDark
-                  ? '0 2px 20px rgba(0,0,0,0.8)'
-                  : '0 1px 8px rgba(255,255,255,0.8)',
-              }}
-            >
-              {movie.name}
-            </h2>
-            <div
-              className="flex items-center gap-3 flex-wrap"
-              style={{ color: 'var(--pc-t2)', fontSize: '0.82rem' }}
-            >
-              <span>{movie.year}</span>
-              {movie.age_rating && (
-                <>
-                  <span>·</span>
-                  <span
-                    className="px-1.5 py-0.5 rounded"
-                    style={{
-                      border: '1px solid var(--pc-bd3)',
-                      color: 'var(--pc-t2)',
-                    }}
-                  >
-                    {movie.age_rating}
-                  </span>
-                </>
-              )}
-              {movie.duration && (
-                <>
-                  <span>·</span>
-                  <span className="flex items-center gap-1">
-                    <Clock size={11} />
-                    {movie.duration} min
-                  </span>
-                </>
-              )}
-              {score > 0 && (
-                <>
-                  <span>·</span>
-                  <span className="flex items-center gap-1">
-                    <Star size={11} fill="#F5C518" stroke="none" />
-                    <span style={{ color: 'var(--pc-gold)' }}>{score}</span>/10
-                  </span>
-                </>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Content */}
-      <div className="p-6 pt-4">
-        {!movie.posterURL && (
-          <div className="mb-4">
-            <h2
-              style={{
-                fontFamily: "var(--font-bebas-neue), 'Bebas Neue', sans-serif",
-                fontSize: '2rem',
-                letterSpacing: '0.04em',
-                color: 'var(--pc-t1)',
-              }}
-            >
-              {movie.name}
-            </h2>
-            <div
-              className="flex items-center gap-3 flex-wrap mt-1"
-              style={{ color: 'var(--pc-t2)', fontSize: '0.82rem' }}
-            >
-              <span>{movie.year}</span>
-              {movie.age_rating && (
-                <>
-                  <span>·</span>
-                  <span>{movie.age_rating}</span>
-                </>
-              )}
-              {movie.duration && (
-                <>
-                  <span>·</span>
-                  <span>{movie.duration} min</span>
-                </>
-              )}
-            </div>
-          </div>
-        )}
-
-        <div className="flex items-center justify-between flex-wrap gap-3 mb-4">
-          <div className="flex items-center gap-2">
-            <Clapperboard size={13} style={{ color: 'var(--pc-t3)' }} />
-            <span style={{ color: 'var(--pc-t3)', fontSize: '0.82rem' }}>
-              {movie.year} · {pct}% match
-            </span>
-          </div>
-          {movie.age_rating && <GenrePill genre={movie.age_rating} />}
-        </div>
-
-        {score > 0 && (
-          <div className="flex items-center gap-2 mb-4">
-            <StarRating score={score} />
-            <span style={{ color: 'var(--pc-t3)', fontSize: '0.78rem' }}>{score}/10</span>
-          </div>
-        )}
-
-        {/* AI description */}
-        {movie.description && (
-          <div
-            className="p-4 rounded-2xl"
-            style={{
-              background: isDark ? 'rgba(245,197,24,0.05)' : 'rgba(196,149,10,0.05)',
-              border: '1px solid',
-              borderColor: isDark ? 'rgba(245,197,24,0.1)' : 'rgba(196,149,10,0.15)',
-            }}
-          >
-            <div className="flex items-center gap-2 mb-2">
-              <Sparkles size={12} style={{ color: 'var(--pc-gold)' }} />
-              <span
-                className="uppercase tracking-widest"
-                style={{ color: 'var(--pc-gold)', fontSize: '0.65rem' }}
-              >
-                Why this film for you
-              </span>
-            </div>
-            <p
-              style={{
-                color: 'var(--pc-t2)',
-                fontSize: '0.88rem',
-                lineHeight: 1.75,
-              }}
-            >
-              {movie.description}
-            </p>
-          </div>
-        )}
-      </div>
-    </motion.div>
-  );
-}
-
-function SmallSuggestionCard({
-  movie,
-  active,
-  onClick,
-}: {
-  movie: MovieRecommendation;
-  active: boolean;
-  onClick: () => void;
-}) {
-  const [imgLoaded, setImgLoaded] = useState(false);
-  const score = movie.score_rating ?? 0;
-  const pct = Math.round(movie.similarity * 100);
-
-  return (
-    <motion.div
-      layout
-      onClick={onClick}
-      className="relative overflow-hidden rounded-2xl cursor-pointer transition-all duration-300"
-      style={{
-        background: 'var(--pc-surface)',
-        border: active ? '1.5px solid rgba(245,197,24,0.4)' : '1px solid var(--pc-bd2)',
-        boxShadow: active ? '0 0 30px rgba(245,197,24,0.1)' : 'none',
-        minWidth: '220px',
-        maxWidth: '260px',
-        flex: '0 0 auto',
-      }}
-      whileHover={{ y: -4 }}
-      whileTap={{ scale: 0.97 }}
-    >
-      <div className="relative h-36 overflow-hidden">
-        {movie.posterURL ? (
-          <>
-            <div
-              className="absolute inset-0"
-              style={{
-                background: 'var(--pc-surface-deep)',
-                opacity: imgLoaded ? 0 : 1,
-              }}
-            />
-            <img
-              src={movie.posterURL}
-              alt={movie.name}
-              className="w-full h-full object-cover"
-              onLoad={() => setImgLoaded(true)}
-              style={{ opacity: imgLoaded ? 1 : 0, transition: 'opacity 0.4s' }}
-            />
-          </>
-        ) : (
-          <div
-            className="w-full h-full flex items-center justify-center text-3xl"
-            style={{ background: 'var(--pc-surface-deep)' }}
-          >
-            🎬
-          </div>
-        )}
-        <div
-          className="absolute inset-0"
-          style={{
-            background: 'linear-gradient(to bottom, transparent 40%, var(--pc-surface) 100%)',
-          }}
-        />
-        <div className="absolute top-2 right-2">
-          <div
-            className="text-xs px-2 py-0.5 rounded-full"
-            style={{
-              background: 'rgba(9,9,15,0.85)',
-              border: '1px solid rgba(245,197,24,0.2)',
-              color: '#F5C518',
-              backdropFilter: 'blur(6px)',
-            }}
-          >
-            {pct}%
-          </div>
-        </div>
-      </div>
-
-      <div className="p-3.5">
-        <h4
-          className="mb-1 truncate"
-          style={{ color: 'var(--pc-t1)', fontWeight: 600, fontSize: '0.9rem' }}
-        >
-          {movie.name}
-        </h4>
-        <div
-          className="flex items-center gap-2 mb-2 flex-wrap"
-          style={{ color: 'var(--pc-t3)', fontSize: '0.75rem' }}
-        >
-          <span>{movie.year}</span>
-          {movie.age_rating && (
-            <>
-              <span>·</span>
-              <span>{movie.age_rating}</span>
-            </>
-          )}
-          {score > 0 && (
-            <>
-              <span>·</span>
-              <span className="flex items-center gap-0.5">
-                <Star size={10} fill="#F5C518" stroke="none" />
-                {score}
-              </span>
-            </>
-          )}
-          {movie.duration && (
-            <>
-              <span>·</span>
-              <span>{movie.duration}m</span>
-            </>
-          )}
-        </div>
-      </div>
-    </motion.div>
-  );
-}
-
-function ExpandedSuggestion({ movie }: { movie: MovieRecommendation }) {
-  const { isDark } = usePCTheme();
-  const score = movie.score_rating ?? 0;
-
-  return (
-    <motion.div
-      initial={{ opacity: 0, height: 0 }}
-      animate={{ opacity: 1, height: 'auto' }}
-      exit={{ opacity: 0, height: 0 }}
-      transition={{ duration: 0.35 }}
-      className="overflow-hidden"
-    >
-      <div
-        className="p-5 rounded-2xl mt-3"
-        style={{
-          background: 'var(--pc-surface)',
-          border: '1px solid',
-          borderColor: isDark ? 'rgba(245,197,24,0.15)' : 'rgba(196,149,10,0.2)',
-        }}
-      >
-        <div className="flex items-start gap-4">
-          {movie.posterURL && (
-            <div className="relative shrink-0 w-20 h-28 rounded-xl overflow-hidden">
-              <img src={movie.posterURL} alt={movie.name} className="w-full h-full object-cover" />
-            </div>
-          )}
-          <div className="flex-1 min-w-0">
-            <div className="flex items-start justify-between gap-2 mb-2">
-              <div>
-                <h3
-                  style={{
-                    fontFamily: "var(--font-bebas-neue), 'Bebas Neue', sans-serif",
-                    fontSize: '1.3rem',
-                    letterSpacing: '0.03em',
-                    color: 'var(--pc-t1)',
-                  }}
-                >
-                  {movie.name}
-                </h3>
-                <div
-                  className="flex items-center gap-2"
-                  style={{ color: 'var(--pc-t3)', fontSize: '0.78rem' }}
-                >
-                  <span>{movie.year}</span>
-                  {movie.age_rating && (
-                    <>
-                      <span>·</span>
-                      <span>{movie.age_rating}</span>
-                    </>
-                  )}
-                  {score > 0 && (
-                    <>
-                      <span>·</span>
-                      <span className="flex items-center gap-0.5">
-                        <Star size={10} fill="#F5C518" stroke="none" />
-                        {score}
-                      </span>
-                    </>
-                  )}
-                </div>
-              </div>
-              <SimilarityBadge similarity={movie.similarity} />
-            </div>
-            {score > 0 && <StarRating score={score} />}
-          </div>
-        </div>
-
-        {movie.description && (
-          <div
-            className="mt-4 p-3.5 rounded-xl"
-            style={{
-              background: isDark ? 'rgba(245,197,24,0.04)' : 'rgba(196,149,10,0.05)',
-              border: '1px solid',
-              borderColor: isDark ? 'rgba(245,197,24,0.08)' : 'rgba(196,149,10,0.12)',
-            }}
-          >
-            <div className="flex items-center gap-2 mb-1.5">
-              <Sparkles size={11} style={{ color: 'var(--pc-gold)' }} />
-              <span
-                className="uppercase tracking-widest"
-                style={{ color: 'var(--pc-gold)', fontSize: '0.62rem' }}
-              >
-                Why this film
-              </span>
-            </div>
-            <p
-              style={{
-                color: 'var(--pc-t2)',
-                fontSize: '0.82rem',
-                lineHeight: 1.7,
-              }}
-            >
-              {movie.description}
-            </p>
-          </div>
-        )}
-      </div>
-    </motion.div>
-  );
 }
 
 export default function ResultsPage() {
