--- src/components/Branding/index.ts
@@ -1,1 +1,0 @@
-export * from './Branding';
