+++ src/components/ProgressDots/index.ts
@@ -1,0 +1,1 @@
+export { ProgressDots } from './ProgressDots';
