--- src/components/QuestionsForm/index.ts
@@ -1,3 +1,0 @@
-export * from './MultipleChoiceQuestion';
-export * from './QuestionCard';
-export * from './QuestionsForm';
