--- src/components/TopNavigation/index.ts
@@ -1,1 +1,0 @@
-export { TopNavigation } from './TopNavigation';
