+++ src/components/index.ts
@@ -1,12 +1,11 @@
 export * from './AgeRatingChip';
-export * from './Branding';
 export * from './Button';
 export * from './Mascot';
 export * from './MovieGrid';
 export * from './MoviesTable';
-export * from './QuestionsForm';
+export * from './PCLayout';
+export * from './ProgressDots';
 export * from './SuggestionCard';
 export * from './ThemeProvider';
 export * from './ThemeToggle';
 export * from './TMDBAttribution';
-export * from './TopNavigation';
